Maintaining one json file
