--- BE/active_room.json
@@ -1,1 +1,0 @@
-[]
