--- BE/active_rooms.json
@@ -1,1 +1,0 @@
-[]
