--- BE/rooms_data.json
@@ -1,150 +1,0 @@
-{
-    "6a21c169-dec1-4803-8f51-cb6ce3aded5c": {
-        "users": [
-            {
-                "userId": "stringddfghjk",
-                "userName": "stringrtyujk"
-            }
-        ]
-    },
-    "b98a4793-ff8f-4def-a779-4fee4c97f858": {
-        "users": [
-            {
-                "userId": "stringhgjj",
-                "userName": "stringfghjktyuy4edf"
-            }
-        ]
-    },
-    "28db8719-b15d-4dd0-b2f8-0baf6bcb71e1": {
-        "users": [
-            {
-                "userId": "stringiuyythjhjj",
-                "userName": "string,nncffd"
-            }
-        ]
-    },
-    "0677e7c9-c51d-4a6d-8488-357c5a8301a5": {
-        "users": [
-            {
-                "userId": "stringmnbbhgfv",
-                "userName": "stringrtyuikjn"
-            }
-        ]
-    },
-    "9294f785-e453-4871-8916-8478bbf10d20": {
-        "users": [
-            {
-                "userId": "strighghng",
-                "userName": "strinbnhghhfg"
-            }
-        ]
-    },
-    "8fd129a6-9cd8-40bb-88a5-e128ea300c06": {
-        "users": [
-            {
-                "userId": "strsfsfing",
-                "userName": "strsffing"
-            }
-        ]
-    },
-    "fdb147c1-ecbe-4ec7-a92a-98f9f226a5e9": {
-        "users": [
-            {
-                "userId": "stsfdgfring",
-                "userName": "strindgfgg"
-            }
-        ]
-    },
-    "93be6dac-29fa-4012-a3d9-be5b10b55deb": {
-        "users": [
-            {
-                "userId": "strinbcvfgg",
-                "userName": "stringfhhghg"
-            }
-        ]
-    },
-    "46d7cb30-0c9a-40fe-92e0-8ffe5900b3b4": {
-        "users": [
-            {
-                "userId": "strindfdgfg",
-                "userName": "strigggbng"
-            }
-        ]
-    },
-    "28b59a39-526b-4033-9819-0a4a026c1e8e": {
-        "users": []
-    },
-    "bdfea2f4-2d78-4a84-8948-d73ae37d9d1c": {
-        "users": []
-    },
-    "aba400d9-0b18-498c-9008-360a9021ac9b": {
-        "users": []
-    },
-    "ac12e398-724a-4e38-8d4c-f60863df5884": {
-        "users": [
-            {
-                "userId": "strixcvcng",
-                "userName": "stringbcv"
-            }
-        ]
-    },
-    "51c68bec-57a9-4285-8b18-7826db8955bb": {
-        "users": [
-            {
-                "userId": "strinfdsg",
-                "userName": "strfdedring"
-            }
-        ]
-    },
-    "24b07849-6072-4875-8599-6d36cd2d3440": {
-        "users": []
-    },
-    "54e7c292-83ed-416c-ba67-143d4d1404f8": {
-        "users": []
-    },
-    "ffb44472-e5c8-4bac-9d0c-27eb68cec578": {
-        "users": [
-            {
-                "userId": "strixcng",
-                "userName": "strinmng"
-            }
-        ]
-    },
-    "ca3140e6-16aa-4f73-a2e1-0e68ea862f15": {
-        "users": []
-    },
-    "f27a28df-a40b-485c-a0f8-45658b5d801f": {
-        "users": []
-    },
-    "fd6ab81e-1229-494d-8d77-f5e7e49cd9fa": {
-        "users": []
-    },
-    "bd4d195e-0105-4934-9968-6523e1cf875e": {
-        "users": [
-            {
-                "userId": "strinffffg",
-                "userName": "stringggg"
-            }
-        ]
-    },
-    "cf04414a-267e-4d73-9f52-2d9c7afb2b0e": {
-        "users": []
-    },
-    "65c7f3b2-1aee-40bb-93ea-f2cec4bf419a": {
-        "users": []
-    },
-    "becdc9f6-e0c2-4571-ad45-c5d1c8df3426": {
-        "users": []
-    },
-    "eb6393f8-0daf-4315-85b7-a97118087916": {
-        "users": [
-            {
-                "userId": "123",
-                "userName": "mohan"
-            }
-        ]
-    },
-    "2ce5eb8a-a522-4c16-b2b9-454361184d9b": {
-        "users": []
-    }
-}
